requirements.txt

aiologger[aiofiles]
pip-chill
gunicorn
fastapi
httpx
pandas
openai
beautifulsoup4
uvicorn
pydantic
python-dotenv
